Move ServiceOrchestrator into runtime::service::instance
The orchestrator should never have been a separate module as
it is solely related to services. This commit moves the orchestrator
into the runtime::service::instance module along side the other
service instance code.

Two traits of the orchestrator module have been moved out to
service::instance module instead because these components are
pulled in by the Scabbard service. The scabbard service should
not pull in anything from the runtime module.
