+++ libsplinter/src/service/instance/mod.rs
@@ -17,6 +17,7 @@ mod factory;
 mod message_context;
 mod network_registry;
 mod network_sender;
+mod orchestrator;
 mod service_instance;
 mod validation;
 
@@ -24,6 +25,7 @@ pub use factory::ServiceFactory;
 pub use message_context::ServiceMessageContext;
 pub use network_registry::ServiceNetworkRegistry;
 pub use network_sender::ServiceNetworkSender;
+pub use orchestrator::{OrchestratableService, OrchestratableServiceFactory};
 pub use service_instance::ServiceInstance;
 pub use validation::ServiceArgValidator;
